Eval was failing, seems to be a bug with eq? vs equal?
I was trying to run some examples in DrRacket and the examples were
failing with "Got Stuck!". It seems the issue was the ast.rkt
substitution. It was failing to substitute variables, changing eq?
to equal? seems to work. Not much of a racket developer, so don't
know if this is the proper check.